Background fonctionnel (avec update) + un peu de cleanup

+++ src/Camera.java
@@ -1,12 +1,12 @@
 public class Camera {
 
-    private int x,y;
+    private long x,y;
 
     public Camera(int xcoord,int ycoord){x = xcoord; y = ycoord;}
 
-    public int getX() {return x;}
+    public long getX() {return x;}
 
-    public int getY() {return y;}
+    public long getY() {return y;}
 
     @Override
     public String toString(){return("x coordinate = " + x + "\r\n" + "y coordinate = " + y);}
